added type casting to progress calculation

+++ DXDAL/Code/Classes/HTTP/DXDALDataProviderHTTP.m
@@ -77,7 +77,7 @@
     
     void (^progressBlock)(NSInteger bytesRead, NSInteger totalBytesRead, NSInteger totalBytesExpectedToRead);
     progressBlock = ^(NSInteger bytesRead, NSInteger totalBytesRead, NSInteger totalBytesExpectedToRead) {
-        float progress = totalBytesExpectedToRead / totalBytesRead;
+        float progress = (float)totalBytesExpectedToRead / (float)totalBytesRead;
         [httpRequest didChangeProgressValue:progress];
     };
     [operation setDownloadProgressBlock:progressBlock];
